+++ seeds/vietinbank/vietinbank.js
@@ -3,7 +3,7 @@ import { loadAllCsvFiles } from "../transaction-seeder.js";
 import path from "path";
 import { fileURLToPath } from "url";
 
-async function vietinbank() {
+async function vietinbank(movingFile = true, deleteAll = false) {
 	const __filename = fileURLToPath(import.meta.url);
 	const __dirname = path.dirname(__filename);
 	const prisma = new PrismaClient();
@@ -11,14 +11,16 @@ async function vietinbank() {
 	prisma.$connect();
 
 	const nameBank = "vietinbank";
-	await prisma.statement.deleteMany({
-		where: {
-			bank_name: nameBank,
-		},
-	});
+	if (deleteAll) {
+		await prisma.statement.deleteMany({
+			where: {
+				bank_name: nameBank,
+			},
+		});
+	}
 
 	// transactionSeeder("./files/bidv_1261122666_09-01_to_09-12-2024.csv", prisma);
-	loadAllCsvFiles(`${__dirname}/files/`, nameBank, prisma);
+	loadAllCsvFiles(`${__dirname}/files/`, nameBank, prisma, movingFile);
 	await prisma.$disconnect();
 }
 export { vietinbank };
